Update Rust Implementation
I wasn't getting it to compile, due to some errors. It is likely caused by the Cargo.toml file not specifying a specific version of euclid, so it is giving errors with the updated version. This should compile now. With gdscript I was able to get 11,713, and in rust I was able to get 24,291if that tells you the performance delta.